move AssignmentKind and CoreAssigment to scheduler
